fix: move `ClientModule` trait into submodule
